[feat] Log boot status only on an error
Originally, boot statuses were logged to the boot status register as the ROM progressed with it's execution.
This potentially opens up a surface for glitching attacks as the exact point of ROM execution is known.
With this change, the boot status is stored  in the DCCM and is updated in the boot status register only
in the case of a failure.